all registers except the IP now print in hex. I left IP printing in base 10 because the .lst file displays it in base 10. Added a quit command. Removed useless member from function Small type change in struct_decl

+++ interp/dbg/struct_decl.h
@@ -21,13 +21,13 @@ public:
 	struct_decl() {}
 
 	string to_string( Machine_State* cpu, uint32_t indent_level = 0 ) const;
-	inline int size() const { return _size; }
+	inline size_t size() const { return _size; }
 	inline string name() const { return _name; }
         variable* var( const string& var_name );
 
 private:
 	vector<variable> _fields;
-	int _size = 0;
+	size_t _size = 0;
 	string _name;
 };
